Moved outdated? to a share space to reduce copy and pasted code.
